Added docs to dataquality module
